Use <pattern> to hold the image tag
This change puts the image tag containing the canvas
render of the mesh gradient in a <pattern> tag in defs,
which replaces the meshGradient while taking the same id
as the meshGradient it's replacing. The image is not directly
inserted in the document.

This way any element referring to the meshGradient, now
automatically refers to the pattern. This allows usage of
meshGradient as stroke as well as fill. Also we don't have
to do any custom clipping in the polyfill code. The SVG
implementation handles it.