Offset for BBox adjustment
The samples so far had the starting point of the mesh gradient patches
to be the xmin,ymin point of bbox. But that easily is not a general
case. sample 4 and 6 are examples of that case.

To correctly render such gradient, we have to do offset adjustment based
on bbox xmin,ymin.